requirements.txt
Pillow>=9.0.0
numpy
opencv-python
pyttsx3
pytest
typer